+++ src/App.tsx
@@ -5,6 +5,7 @@ import Home from './pages/home/home.component';
 import FearGreedIndex from './components/fear-greed-index/fear-greed-index.component';
 import Navbar from './components/navbar/navbar.component';
 import Topbar from './components/topbar/topbar.component';
+import Market from './pages/market/market.component';
 
 function App() {
 	return (
@@ -12,6 +13,7 @@ function App() {
 			<Topbar />
 			<Navbar />
 			<Home />
+			<Market />
 			<FearGreedIndex />
 		</div>
 	);